dont store duplicated entries in json file

+++ src/encodeURL.py
@@ -1,11 +1,13 @@
 import hashlib
 import json
 
-jsonPath = "../../data/links.json"
+jsonPath = "../data/links.json"
 
 def storeUrl(url, shortened):
     with open(jsonPath, 'r') as file:
         json_data = json.load(file)
+    if json_data[shortened]: #if shortened sha is already present, exit the function
+        return
     json_data[shortened] = url
     with open(jsonPath, 'w') as file:
         json.dump(json_data, file, indent=4)
